requirements.txt
Flask==0.10.1
Flask-PyMongo==0.4.1
gunicorn==19.4.5
pymongo==3.2.2
python-dateutil==2.5.1
